Rename CanLookup RSpec shared example group to follow existing pattern
