+++ spec/models/concerns/can_lookup_examples.rb
