+++ spec/models/organization_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require 'models/concerns/can_lookup'
+require 'models/concerns/can_lookup_examples'
 
 RSpec.describe Organization do
   include_examples 'CanLookup'
